// model/ethereumWalletAddress.ts
/**
 * Bleumi Pay API
 * A simple and powerful REST API to integrate Algorand, Ethereum, ERC-20 and xDai payments and/or payouts into your business
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@bleumi.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { EthereumWalletInputs } from './ethereumWalletInputs';

export class EthereumWalletAddress {
    /**
    * Wallet address for the payment in the network
    */
    'addr': string;
    'inputs'?: EthereumWalletInputs;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "addr",
            "baseName": "addr",
            "type": "string"
        },
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "EthereumWalletInputs"
        }    ];

    static getAttributeTypeMap() {
        return EthereumWalletAddress.attributeTypeMap;
    }
}