// model/payment.ts
/**
 * Bleumi Pay API
 * A simple and powerful REST API to integrate ERC-20, Ethereum, xDai payments and/or payouts into your business or application
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@bleumi.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { PaymentAddresses } from './paymentAddresses';
import { PaymentBalances } from './paymentBalances';

export class Payment {
    /**
    * Unique ID identifying the payment; specified when it was created by your system
    */
    'id': string;
    'addresses': PaymentAddresses;
    'balances': PaymentBalances;
    /**
    * UNIX timestamp when the payment was created
    */
    'createdAt': number;
    /**
    * UNIX timestamp when the lastest operation was performed
    */
    'updatedAt': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "addresses",
            "baseName": "addresses",
            "type": "PaymentAddresses"
        },
        {
            "name": "balances",
            "baseName": "balances",
            "type": "PaymentBalances"
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "number"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return Payment.attributeTypeMap;
    }
}